move helper methods to HelperTrait
