+++ src/Consumer.php
@@ -4,6 +4,7 @@ namespace Vinelab\Bowler;
 
 use PhpAmqpLib\Message\AMQPMessage;
 use Vinelab\Bowler\Traits\AdminTrait;
+use Vinelab\Bowler\Traits\HelperTrait;
 use Vinelab\Bowler\Traits\DeadLetteringTrait;
 use Vinelab\Bowler\Exceptions\DeclarationMismatchException;
 use Vinelab\Bowler\Contracts\BowlerExceptionHandler as ExceptionHandler;
@@ -17,6 +18,7 @@ use Vinelab\Bowler\Contracts\BowlerExceptionHandler as ExceptionHandler;
 class Consumer
 {
     use AdminTrait;
+    use HelperTrait;
     use DeadLetteringTrait;
 
     /**
